requirements.txt
streamlit==1.29.0
tensorflow==2.13.0
numpy==1.25.0
Pillow==10.1.0
gdown==4.7.1
